+++ src/main/java/core/basesyntax/Machine.java
@@ -1,4 +1,4 @@
-package OOPBasic;
+package core.basesyntax;
 
 public abstract class Machine {
     abstract void doWork();
